usize on one

+++ sequencer/src/lib.rs
@@ -142,7 +142,7 @@ pub struct NetworkParams {
     pub libp2p_mesh_outbound_min: usize,
 
     /// The maximum gossip message size
-    pub libp2p_max_gossip_transmit_size: u64,
+    pub libp2p_max_gossip_transmit_size: usize,
 
     /// The maximum direct message size
     pub libp2p_max_direct_transmit_size: u64,
